数据字典分页查询成功

+++ projects/TeachingManagement/SystemManagement/src/main/java/com/ruicai/system/vo/Datagrid.java
@@ -9,14 +9,14 @@ import java.util.List;
  * Created by lws on 2017/4/11.
  */
 public class Datagrid<T> {
-    private int total;
+    private long total;
     private List<T> rows = new ArrayList<>();
 
-    public int getTotal() {
+    public long getTotal() {
         return total;
     }
 
-    public void setTotal(int total) {
+    public void setTotal(long total) {
         this.total = total;
     }
